change our key/id attribute names to be more consistent with AWS
...and retain compatibility with "id" and "key"; both as accessors and legal
parameters to pass to new().

HOWEVER: MultiInitArg seems to be barfing on this for some reason.  It might
be that my brain is just a little fried right now and I'm not seeing something
obvious, as it's tests pass and I can't see any obvious metaclass magic
mishaps that could explain this.